apps/api/requirements.txt
asgiref==3.11.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
dj-database-url==3.1.0
Django==5.2.9
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.45.0
gunicorn==23.0.0
idna==3.11
packaging==25.0
pillow==12.1.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
python-dotenv==1.2.1
qrcode==8.2
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.2
whitenoise==6.11.0
